fix: P2 trivials — LDLT->Cholesky doc + explicit cast + dead constant
Three trivial P2 fixes:

1. src/calib_linalg.hpp:39 — doc comment said 'LDLT factorization' but
   impl is plain Cholesky (LL^T) via dpotrf, no D pivoting. Updated.

2. src/calib_validate.cpp:38-39 — explicit static_cast<double>() on int
   n_per_cell[c] before multiply with double. Silences -Wconversion;
   no behavior change.

3. src/ieppa.cpp:452 — kInfeasStallRatio defined once in
   ieppa_finalize() at line 148; ieppa_solve()'s redefinition was
   dead. Removed; comment now points to single source of truth.

Closes leafblower-z3gk (T29), leafblower-dig0 (T44), leafblower-y51s (T45).

+++ src/calib_validate.cpp
@@ -35,8 +35,8 @@ int calib_validate_preentry(const CellTable& ct,
     double sum_L = 0.0, sum_U = 0.0;
 
     for (int c = 0; c < ct.M_cell; c++) {
-        double L_c = lo * ct.n_per_cell[c];
-        double U_c = hi * ct.n_per_cell[c];
+        double L_c = lo * static_cast<double>(ct.n_per_cell[c]);
+        double U_c = hi * static_cast<double>(ct.n_per_cell[c]);
 
         // 2a. Empty cell guard
         if (ct.n_per_cell[c] == 0) {
